langchain_krishnaik_6projects/requirements.txt
langchain
openai
python-dotenv
streamlit
PyPDF2
dataset